+++ src/support/utils.js
@@ -11,9 +11,10 @@
  */
 
 import { context as h2, h1 } from '@adobe/fetch';
-import { hasText } from '@adobe/spacecat-shared-utils';
+import { hasText, resolveCustomerSecretsName } from '@adobe/spacecat-shared-utils';
 import URI from 'urijs';
 import { JSDOM } from 'jsdom';
+import { GetSecretValueCommand, SecretsManagerClient } from '@aws-sdk/client-secrets-manager';
 
 URI.preventInvalidHostname = true;
 
@@ -141,4 +142,28 @@ export function getSitemapUrlsFromSitemapIndex(content) {
 export function getUrlWithoutPath(url) {
   const urlObj = new URL(url);
   return `${urlObj.protocol}//${urlObj.host}`;
+}
+
+/**
+ * Retrieves the RUM domain key for the specified base URL from the customer secrets.
+ *
+ * @param {string} baseURL - The base URL for which the RUM domain key is to be retrieved.
+ * @param {UniversalContext} context - Helix Universal Context. See https://github.com/adobe/helix-universal/blob/main/src/adapter.d.ts#L120
+ * @returns {Promise<string>} - A promise that resolves to the RUM domain key.
+ * @throws {Error} Throws an error if no domain key is found for the specified base URL.
+ */
+export async function getRUMDomainkey(baseURL, context) {
+  const customerSecretName = resolveCustomerSecretsName(baseURL, context);
+  const { runtime } = context;
+
+  try {
+    const client = new SecretsManagerClient({ region: runtime.region });
+    const command = new GetSecretValueCommand({
+      SecretId: customerSecretName,
+    });
+    const response = await client.send(command);
+    return JSON.parse(response.SecretString)?.RUM_DOMAIN_KEY;
+  } catch (error) {
+    throw new Error(`Error retrieving the domain key for ${baseURL}. Error: ${error.message}`);
+  }
 }